Update to Substrate master and Parachains V1 (#102)
Overview

This PR is primarily an update of four key Moonbeam dependencies, Substrate, Cumulus, Polkadot, and Frontier. For a while Moonbeam has been using old versions of Substrate, Cumulus, and Polkadot, and backporting Frontier for compatability. This PR updates to the master branch of all of those.1
New Features

Although no new features are added directly to moonbeam with this upgrade, we gain several new features or fixes from upstream.

    Parachains v1 (Polkadot)
    Full multi collator support (Cumulus)
    Weight gas mapping (Frontier)
    More rpc endpoints (Frontier)
    Rename trait Trait to trait Config (Substrate)
    Estimate gas from EVM (Frontier)

Hotfix Status

Two hotfixes were removed. We no longer use a patched Frontier runtime API. Because of the Substrate update, we no longer use a non-standard ref_count type.

Two hotfixes remain, but were moved to Moonbeam as drop-in replacements for their upstream counterparts. This allows us to depend on the stock Frontier, and keep better track of hotfixes because they live in the Moonbeam repo. Specifically the EthApi was moved.This includes two orthogonal patches:

    The backend optimization
    The headers patch